Add TxProcessor structure
