+++ crates/indexer/src/indexer/mod.rs
@@ -2,10 +2,12 @@ mod block_processor;
 mod cache;
 mod db;
 mod handlers;
+mod tx_processor;
 pub mod worker;
 
 pub use block_processor::*;
 pub use cache::UtxoCache;
 pub use db::*;
 pub use handlers::*;
+pub use tx_processor::*;
 pub use worker::Worker;
